Unify chat engines into single UnifiedChatEngine with enhanced capabilities
