+++ app/main.py
@@ -22,9 +22,8 @@ from app.chat_models import (
     MemoryUpdate
 )
 from app.core.research_engine import ResearchEngine
-from app.core.chat_engine import ChatEngine
+from app.core.unified_chat_engine import UnifiedChatEngine
 from app.websocket_handler import websocket_endpoint
-from app.core.enhanced_chat_engine import EnhancedChatEngine
 
 # Configure logging
 structlog.configure(
@@ -48,8 +47,7 @@ logger = structlog.get_logger()
 # Store research briefs in memory (use Redis/DB in production)
 research_store: Dict[str, ResearchBrief] = {}
 research_engines: Dict[str, ResearchEngine] = {}
-chat_engine = ChatEngine()  # Global chat engine instance
-enhanced_engine = EnhancedChatEngine()  # Enhanced conversational engine
+chat_engine = UnifiedChatEngine()  # Unified chat engine instance
 
 
 @asynccontextmanager
